working tweaks
